BroadleafCommerce/QA#4053 Clear Offer qualifiers when switching offer type
