+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/offer/service/AbstractOfferServiceExtensionHandler.java
@@ -27,7 +27,6 @@ import org.broadleafcommerce.core.offer.service.discount.domain.PromotableOrder;
 import org.broadleafcommerce.core.offer.service.discount.domain.PromotableOrderItem;
 import org.broadleafcommerce.core.order.domain.Order;
 import org.broadleafcommerce.core.order.domain.OrderItemPriceDetail;
-import org.broadleafcommerce.openadmin.dto.Entity;
 import org.broadleafcommerce.profile.core.domain.Customer;
 
 import java.util.HashMap;
@@ -93,9 +92,6 @@ public class AbstractOfferServiceExtensionHandler extends AbstractExtensionHandl
         return ExtensionResultStatusType.NOT_HANDLED;
     }
 
-    @Override
-    public  ExtensionResultStatusType clearHiddenQualifiers(Entity entity) {
-        return ExtensionResultStatusType.NOT_HANDLED;
-    }
+
 
 }
